backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Bcrypt
Flask-CORS
spacy
transformers
torch
tensorflow
scikit-learn
pandas
numpy
nltk
textblob
joblib
xgboost
python-dotenv
werkzeug
email-validator
config